Extract entity template functions into an entity Jinja2 extension
