+++ tests/helpers/template/test_init.py
@@ -26,12 +26,7 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import TemplateError
-from homeassistant.helpers import (
-    device_registry as dr,
-    entity_registry as er,
-    template,
-    translation,
-)
+from homeassistant.helpers import entity_registry as er, template, translation
 from homeassistant.helpers.json import json_dumps
 from homeassistant.helpers.template.render_info import (
     ALL_STATES_RATE_LIMIT,
@@ -401,85 +396,6 @@ def test_if_state_exists(hass: HomeAssistant) -> None:
         hass, "{% if states.test.object %}exists{% else %}not exists{% endif %}"
     )
     assert result == "exists"
-
-
-def test_entity_name(
-    hass: HomeAssistant,
-    entity_registry: er.EntityRegistry,
-    device_registry: dr.DeviceRegistry,
-) -> None:
-    """Test entity_name method."""
-    assert render(hass, "{{ entity_name('sensor.fake') }}") is None
-
-    entry = entity_registry.async_get_or_create(
-        "sensor", "test", "unique_1", original_name="Registry Sensor"
-    )
-    assert render(hass, f"{{{{ entity_name('{entry.entity_id}') }}}}") == (
-        "Registry Sensor"
-    )
-    assert render(hass, f"{{{{ '{entry.entity_id}' | entity_name }}}}") == (
-        "Registry Sensor"
-    )
-
-    entity_registry.async_update_entity(entry.entity_id, name="My Custom Sensor")
-    assert render(hass, f"{{{{ entity_name('{entry.entity_id}') }}}}") == (
-        "My Custom Sensor"
-    )
-
-    # Falls back to state for entities not in the registry
-    hass.states.async_set(
-        "light.no_unique_id", "on", {"friendly_name": "No Unique ID Light"}
-    )
-    assert render(hass, "{{ entity_name('light.no_unique_id') }}") == (
-        "No Unique ID Light"
-    )
-
-    config_entry = MockConfigEntry(domain="test")
-    config_entry.add_to_hass(hass)
-    device_entry = device_registry.async_get_or_create(
-        config_entry_id=config_entry.entry_id,
-        connections={(dr.CONNECTION_NETWORK_MAC, "12:34:56:AB:CD:EF")},
-        name="My Device",
-    )
-    entry2 = entity_registry.async_get_or_create(
-        "sensor",
-        "test",
-        "unique_2",
-        config_entry=config_entry,
-        device_id=device_entry.id,
-        has_entity_name=True,
-        original_name="Temperature",
-    )
-    assert render(hass, f"{{{{ entity_name('{entry2.entity_id}') }}}}") == (
-        "Temperature"
-    )
-
-    # Strips device name prefix
-    entity_registry.async_update_entity(
-        entry2.entity_id, name="My Device Custom Sensor"
-    )
-    assert render(hass, f"{{{{ entity_name('{entry2.entity_id}') }}}}") == (
-        "Custom Sensor"
-    )
-
-
-def test_is_hidden_entity(
-    hass: HomeAssistant,
-    entity_registry: er.EntityRegistry,
-) -> None:
-    """Test is_hidden_entity method."""
-    hidden_entity = entity_registry.async_get_or_create(
-        "sensor", "mock", "hidden", hidden_by=er.RegistryEntryHider.USER
-    )
-    visible_entity = entity_registry.async_get_or_create("sensor", "mock", "visible")
-    assert render(hass, f"{{{{ is_hidden_entity('{hidden_entity.entity_id}') }}}}")
-
-    assert not render(hass, f"{{{{ is_hidden_entity('{visible_entity.entity_id}') }}}}")
-
-    assert not render(
-        hass,
-        f"{{{{ ['{visible_entity.entity_id}'] | select('is_hidden_entity') | first }}}}",
-    )
 
 
 def test_is_state(hass: HomeAssistant) -> None:
